+++ rand_core/src/impls.rs
@@ -160,60 +160,6 @@ pub fn next_u64_via_fill<R: RngCore + ?Sized>(rng: &mut R) -> u64 {
     u64::from_le_bytes(buf)
 }
 
-/// Implement [`TryRngCore`][crate::TryRngCore] for a type implementing [`RngCore`].
-///
-/// Ideally, `rand_core` would define blanket impls for this, but they conflict with blanket impls
-/// for `&mut R` and `Box<R>`, so until specialziation is stabilized, implementer crates
-/// have to implement `TryRngCore` directly.
-#[macro_export]
-macro_rules! impl_try_rng_from_rng_core {
-    ($t:ty) => {
-        impl $crate::TryRngCore for $t {
-            type Error = core::convert::Infallible;
-
-            #[inline]
-            fn try_next_u32(&mut self) -> Result<u32, Self::Error> {
-                Ok($crate::RngCore::next_u32(self))
-            }
-
-            #[inline]
-            fn try_next_u64(&mut self) -> Result<u64, Self::Error> {
-                Ok($crate::RngCore::next_u64(self))
-            }
-
-            #[inline]
-            fn try_fill_bytes(&mut self, dst: &mut [u8]) -> Result<(), Self::Error> {
-                $crate::RngCore::fill_bytes(self, dst);
-                Ok(())
-            }
-        }
-    };
-}
-
-/// Implement [`TryCryptoRng`] and [`TryRngCore`] for a type implementing [`CryptoRng`].
-///
-/// Ideally, `rand_core` would define blanket impls for this, but they conflict with blanket impls
-/// for `&mut R` and `Box<R>`, so until specialziation is stabilized, implementer crates
-/// have to implement `TryRngCore` and `TryCryptoRng` directly.
-///
-/// [`TryCryptoRng`]: crate::TryCryptoRng
-/// [`TryRngCore`]: crate::TryRngCore
-/// [`CryptoRng`]: crate::CryptoRng
-#[macro_export]
-macro_rules! impl_try_crypto_rng_from_crypto_rng {
-    ($t:ty) => {
-        $crate::impl_try_rng_from_rng_core!($t);
-
-        impl $crate::TryCryptoRng for $t {}
-
-        /// Check at compile time that `$t` implements `CryptoRng`
-        const _: () = {
-            const fn check_crypto_rng_impl<T: $crate::CryptoRng>() {}
-            check_crypto_rng_impl::<$t>();
-        };
-    };
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
